+++ SIMPLVtkLib/SIMPLBridge/VSTriangleGeom.cpp
@@ -124,7 +124,7 @@ void VSTriangleGeom::GetCellPoints(vtkIdType cellId, vtkIdList* ptIds)
 {
   const int numVerts = 3;
 
-  int64_t verts[numVerts];
+  size_t verts[numVerts];
   m_Geom->getVertsAtTri(cellId, verts);
 
   ptIds->SetNumberOfIds(numVerts);
@@ -141,7 +141,7 @@ void VSTriangleGeom::GetPointCells(vtkIdType ptId, vtkIdList* cellIds)
 {
   ElementDynamicList::Pointer elementsContainingList = m_Geom->getElementsContainingVert();
 
-  DynamicListArray<uint16_t, int64_t>::ElementList listArray = elementsContainingList->getElementList(ptId);
+  DynamicListArray<uint16_t, size_t>::ElementList listArray = elementsContainingList->getElementList(ptId);
 
   cellIds->SetNumberOfIds(listArray.ncells);
   for(int i = 0; i < listArray.ncells; i++)
